app engine use-game-engine: pull callback into realtime subscribe
The useChannel abstraction wasn't useful because we only call it once. Put the
callback back inside the useEffect hook to prevent unnecessary re-renders.
